refactor(network): unify LISH detail view across search and peer detail
